c++20 upgrade attempt 2, patched png++ lib again

+++ png++-0.2.9/convert_color_space.hpp
@@ -129,9 +129,9 @@ namespace png
             template< class reader >
             static void handle_alpha(reader& io, uint_32 filler)
             {
-                bool src_alpha = (io.get_color_type() & color_mask_alpha);
+                bool src_alpha = (static_cast<int>(io.get_color_type()) & static_cast<int>(color_mask_alpha));
                 bool src_tRNS = io.has_chunk(chunk_tRNS);
-                bool dst_alpha = traits::get_color_type() & color_mask_alpha;
+                bool dst_alpha = static_cast<int>(traits::get_color_type()) & static_cast<int>(color_mask_alpha);
                 if ((src_alpha || src_tRNS) && !dst_alpha)
                 {
 #ifdef PNG_READ_STRIP_ALPHA_SUPPORTED
@@ -198,7 +198,7 @@ namespace png
             {
                 bool src_rgb =
                     io.get_color_type() & (color_mask_rgb | color_mask_palette);
-                bool dst_rgb = traits::get_color_type() & color_mask_rgb;
+                bool dst_rgb = static_cast<int>(traits::get_color_type()) & static_cast<int>(color_mask_rgb);
                 if (src_rgb && !dst_rgb)
                 {
 #ifdef PNG_READ_RGB_TO_GRAY_SUPPORTED
